+++ src/Castle.Facilities.NHibernateIntegration/SessionStores/CallContextSessionStore.cs
@@ -17,11 +17,11 @@
 namespace Castle.Facilities.NHibernateIntegration.SessionStores
 {
     using System;
-    using System.Collections;
+    using System.Collections.Generic;
+
 #if NETFRAMEWORK
     using System.Runtime.Remoting.Messaging;
 #endif
-
 #if NETFRAMEWORK
     /// <summary>
     /// An implementation of <see cref="ISessionStore" />
@@ -33,21 +33,21 @@ namespace Castle.Facilities.NHibernateIntegration.SessionStores
     /// which relies on .NET Framework CallContext.
     /// </summary>
     /// <exception cref="PlatformNotSupportedException">
-    /// This is not supported in .NET.
+    /// This is not supported anymore in .NET.
     /// </exception>
 #endif
     public class CallContextSessionStore : AbstractDictionaryStackSessionStore
     {
-        protected override IDictionary GetSessionDictionary()
+        protected override IDictionary<string, Stack<SessionDelegate>> GetSessionDictionary()
         {
 #if NETFRAMEWORK
-            return CallContext.GetData(SessionSlotKey) as IDictionary;
+            return CallContext.GetData(SessionSlotKey) as IDictionary<string, Stack<SessionDelegate>>;
 #else
             throw new PlatformNotSupportedException();
 #endif
         }
 
-        protected override void StoreSessionDictionary(IDictionary dictionary)
+        protected override void StoreSessionDictionary(IDictionary<string, Stack<SessionDelegate>> dictionary)
         {
 #if NETFRAMEWORK
             CallContext.SetData(SessionSlotKey, dictionary);
@@ -56,16 +56,16 @@ namespace Castle.Facilities.NHibernateIntegration.SessionStores
 #endif
         }
 
-        protected override IDictionary GetStatelessSessionDictionary()
+        protected override IDictionary<string, Stack<StatelessSessionDelegate>> GetStatelessSessionDictionary()
         {
 #if NETFRAMEWORK
-            return CallContext.GetData(StatelessSessionSlotKey) as IDictionary;
+            return CallContext.GetData(StatelessSessionSlotKey) as IDictionary<string, Stack<StatelessSessionDelegate>>;
 #else
             throw new PlatformNotSupportedException();
 #endif
         }
 
-        protected override void StoreStatelessSessionDictionary(IDictionary dictionary)
+        protected override void StoreStatelessSessionDictionary(IDictionary<string, Stack<StatelessSessionDelegate>> dictionary)
         {
 #if NETFRAMEWORK
             CallContext.SetData(StatelessSessionSlotKey, dictionary);